SQL

Vue : I_BENF_ADR_CM
Différentes informations se rapportant aux adresses d'une personne assurée (bénéficiaire) à travers le temps.

SQL_adr(
  user,
  t1 = NULL,
  t2 = NULL,
  cols = c("BENF_NO_INDIV_BEN_BANLS", "LGEO_COD_TERRI_CLSC", "LGEO_COD_TERRI_RLS",
    "LGEO_COD_RSS", "BENF_DD_ADR_BEN", "BENF_DF_ADR_BEN"),
  cols_name = list(BENF_NO_INDIV_BEN_BANLS = "ID", LGEO_COD_TERRI_CLSC = "CLSC",
    LGEO_COD_TERRI_RLS = "RLS", LGEO_COD_RSS = "RSS", BENF_DD_ADR_BEN = "DateDebut",
    BENF_DF_ADR_BEN = "DateFin"),
  cols_type = list(BENF_NO_INDIV_BEN_BANLS = "int", LGEO_COD_TERRI_CLSC = "int",
    LGEO_COD_TERRI_RLS = "int", LGEO_COD_RSS = "int", BENF_DD_ADR_BEN = "date",
    BENF_DF_ADR_BEN = "date"),
  select_ids = NULL,
  save_file = NULL,
  save_ext = "rds",
  ...
)

Arguments

user
Nom d'usager SQL.

t1, t2
Dates au format "AAAA-MM-JJ".
t1=NULL & t2=NULL : Toutes les observations.
t1='AAAA-MM-JJ' & t2=NULL : Observations où la date de début de la période d'application est plus petite ou égale à t1 (BENF_DD_ADR_BEN <= t1).
t1='AAAA-MM-JJ & t2='AAAA-MM-JJ : Observations qui chevauchent l'intervalle [t1, t2] (BENF_DD_ADR_BEN<=t2 et BENF_DF_ADR_BEN>=t1).

cols
Vecteur des colonnes à extraire. NULL implique toutes les colonnes. Voir Details.

cols_name
list(nom_initial = nouveau_nom). Permet de modifier le nom des variables. NULL implique que l'on conserve les noms initiaux.

cols_type
list(nom_initial = type). Permet de modifier/forcer la classe des variables. "int"=integer, "num"=numeric, "chr"=character, "date"=Date et "logical"=logical. NULL implique aucune modification, conservation des classes initiales.

select_ids
Vecteur contenant les bénéficiaires/usagers/individus à conserver.

save_file
"repertoire/nom_fichier" permettant d'enregistrer le dataset sur le disque dur. Exemple : "dir1/dir2/dir3/nom_du_fichier".

save_ext
Extension du fichier à enregistrer. "rds", "csv" ou "txt".

Details

Variables disponibles :
BENF_COD_TYP_ADR : CODE TYPE ADRESSE. Ce code indique à quelle fin sert l'adresse.
BENF_DD_ADR_BEN : Date de début de la période d'application.
BENF_DF_ADR_BEN : Date de fin de la période d'application.
BENF_DM_ADR_BEN : DATE MISE À JOUR ADRESSE BÉNÉFICIAIRE. Date de mise à jour de l'occurrence.
BENF_IND_ADR_HQ : INDICATEUR ADRESSE HORS-QUÉBEC. Ce champ indique si l'adresse inscrite est une adresse qui est hors du Québec (autre province ou hors du Canada).
BENF_NO_INDIV_BEN_BANLS : Numéro identifiant de manière unique une personne physique qui peut être un bénéficiaire d'un service. Ce numéro est banalisé spécifiquement pour le Conseil du médicament.
BENF_NO_INDIV_BEN_PORTR_BANLS : Numéro identifiant de manière unique une personne physique qui peut être un bénéficiaire et porteur de l'adresse d'un autre bénéficiaire. Ce numéro est banalisé spécifiquement pour le Conseil du médicament.
LGEO_COD_RSS : Code de région socio-sanitaire.
LGEO_COD_POSTL_PART : Code postal partiel, c'est-à-dire les 3 premiers caractères à partir de la gauche du code postal.
LGEO_COD_TERRI_RLS : Code de territoire de RLS.
LGEO_COD_TERRI_CLSC : Code de territoire de CLSC
LGEO_COD_TERRI_RUIS : Code de territoire de RUIS
LGEO_COD_LOCAL : CODE LOCALITE.
LGEO_COD_MUNIC : Code de municipalité.
LGEO_COD_MRC : CODE MUNICIPALITES REGIONALES DE COMTE.
LGEO_COD_AIRE_DIFSN_STCAN : Code identifiant une aire de diffusion selon Statistique Canada.